adapt copy function to tests

+++ pyconf/copy.py
@@ -1,30 +1,33 @@
 import os
 import pathlib
 import shutil
+from pathlib import Path
 from shutil import SameFileError
 
 
-def copy(source: str | pathlib.Path, destination: str | pathlib.Path, overwrite: bool = None):
+def copy(
+    source: str | pathlib.Path, destination: str | pathlib.Path, overwrite: bool = False
+):
     """
     copy a resource (file or directory)
     """
     if isinstance(source, str):
-        source = pathlib.Path(source)
+        source: Path = pathlib.Path(source)
     if isinstance(destination, str):
-        destination = pathlib.Path(destination)
-    if destination.exists():
-        if overwrite:
-            try:
-                os.remove(destination)
-            except IsADirectoryError:
-                os.rmdir(destination)
+        destination: Path = pathlib.Path(destination)
+    if destination.exists() and overwrite:
+        try:
+            os.remove(destination)
+        except IsADirectoryError:
+            shutil.rmtree(destination)
+    if destination.exists() and not destination.is_dir() and not overwrite:
+        raise OSError("destination already exists, and overwrite has not been chosen")
     if source.is_dir():
         shutil.copytree(source, destination)
-
-
     elif source.is_file():
         try:
             shutil.copy(source, destination)
         except SameFileError:
             pass
-
+    else:
+        raise OSError("source must be either file or directory")
